# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Ticketing API",
        "description": "API untuk mengelola berbagai resource",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/provinces": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Get all provinces",
                "operationId": "85624171910502f5871da8213efcc4c7",
                "responses": {
                    "200": {
                        "description": "List of provinces",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Province"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Create a new province",
                "operationId": "fbe4a3fda866847e1a434824ff77a951",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "no_province",
                                    "province_name"
                                ],
                                "properties": {
                                    "no_province": {
                                        "type": "string",
                                        "example": "32"
                                    },
                                    "province_name": {
                                        "type": "string",
                                        "example": "Jawa Barat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Province created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Province"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provinces/{id}": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Get a province by ID",
                "operationId": "544a7c0551597906e1f54df1ad064cd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Province details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Province"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Province not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Update a province",
                "operationId": "a43b6c9ad0ed88f568856531cbeded4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "no_province",
                                    "province_name"
                                ],
                                "properties": {
                                    "no_province": {
                                        "type": "string",
                                        "example": "32"
                                    },
                                    "province_name": {
                                        "type": "string",
                                        "example": "Jawa Timur"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Province updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Province"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Province not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Delete a province",
                "operationId": "6b7747c282a844001b0e57990e0d9f45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Province deleted successfully"
                    },
                    "404": {
                        "description": "Province not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provinces/export-format": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Export Format province data",
                "operationId": "4635b90bf77eb19b8dbe0c2f4730d71c",
                "responses": {
                    "200": {
                        "description": "Excel format file exported"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provinces/export": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Export province data",
                "operationId": "30b1bc7a4ddd8ec40fa64d1f98ae56c2",
                "responses": {
                    "200": {
                        "description": "Excel file exported"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provinces/import": {
            "post": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Import province data from Excel",
                "operationId": "e722c682f8f08cfc24b05c799bec6785",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Province": {
                "required": [
                    "no_province",
                    "province_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "no_province": {
                        "type": "string",
                        "example": "32"
                    },
                    "province_name": {
                        "type": "string",
                        "example": "Jawa Barat"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}